Moved loaders into folders
