+++ src/index.ts
@@ -1,20 +1,12 @@
-require('dotenv').config()
-// Require Discord.JS
-import { Client, GatewayIntentBits, Collection } from "discord.js";
-import { commands } from './commands';
-import { events } from './events';
-import { deployCommands } from "./deployCommands";
-const token = process.env.TOKEN;
+import { Client, GatewayIntentBits } from "discord.js";
+import { load_event_handlers } from './events';
+import { refresh_slash_commands } from './commands';
+import { token } from './config';
 
 const client = new Client({ intents: [GatewayIntentBits.Guilds] });
 
-for (const event of events) {
-    if (event.once) {
-        client.once(event.name, (...args) => event.execute(...args));
-    } else {
-        client.on(event.name, (...args) => event.execute(...args));
-    }
-}
-
-
-deployCommands().then(() => client.login(token));
+(async () => {
+    await refresh_slash_commands()
+    load_event_handlers(client)
+    client.login(token)
+})();
